+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * Add first class support for Babashka (more warnings when you connect to `babashka.nrepl`).
+* Add support for nREPL 0.8's `lookup` op.
 
 ### Changes